requirements.txt
alembic==1.4.2
audioread==2.1.8
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
librosa==0.7.2
llvmlite==0.33.0
Mako==1.1.3
MarkupSafe==1.1.1
numba==0.50.1
numpy==1.19.0
paypalrestsdk==1.13.1
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.18
threadpoolctl==2.1.0
twilio==6.44.1
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
